core: Spoof build properties
SafetyNet's CTS profile attestation checks whether Build.MODEL matches
that of the device's stock OS and checking whether it's a device that
matched the CTS Profile.

Spoof the model by setting Build.MODEL to "Pixel 4a (5G)" for
Google Play Services to help pass SafetyNet.

Also while on it, add a model check for Genshin Impact internal testing
tasks that we conduct by setting the Build.MODEL to "Pixel 5 Pro" if
ro.boot.hardware is "gourami" to make sure it not clash with other
devices that's not supposed to be enrolled.

REF: https://github.com/ProtonAOSP/android_frameworks_base/commit/b8adfb5d711c0381350901a40a693b0a85fb9d76

Signed-off-by: spezi77 <spezi7713@gmx.net>
Change-Id: I5b238b61d0bdda47b86b441d93e904ae0df45b01