Remove getProcessGroup usage in OomAdjuster
The returned value may not be the same as it was set. The invocation
was used to verify whether the schedule group is set by zygote. Since
most of time it is success and it is no harm to ignore the result
because the latest state can still be updated from oom-adj computation,
the check of getProcessGroup can be simply removed.

Bug: 211000005
Test: Record a trace during cold launching a top activity.
      The main thread priority should be 100 after attachApplication.
      And "adb shell cat /proc/$pid/task/$pid/cgroup" is top-app.
Change-Id: I282e124d8dc603efdef9b845e8ee6bb61585a6df
Merged-In: I282e124d8dc603efdef9b845e8ee6bb61585a6df
Signed-off-by: spezi77 <spezi7713@gmx.net>